data_eng/weather_streamming_pipeline/requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3